feat: consolidate input and input group components
