+++ components/interactables/Input/types.d.ts
@@ -1,9 +1,10 @@
-export type InputTypes = 'text' | 'number' | 'password'
-export type InputStyle =
+export type InputType = 'text' | 'number' | 'password'
+export type InputColor =
   | 'primary'
   | 'link'
   | 'info'
   | 'success'
   | 'warning'
   | 'danger'
-export type InputSize = 'small' | 'normal' | 'medium' | 'large'
+  | null
+export type InputSize = 'sm' | 'md' | 'lg' | 'xl'
