--- components/interactables/InputGroup/InputGroup.less
@@ -1,55 +1,0 @@
-.input-group {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-
-  .input-outer {
-    display: flex;
-    align-items: stretch;
-    flex-grow: 1;
-    &:extend(.bordered);
-    &:extend(.round-corners);
-
-    .input-inner {
-      position: relative;
-      display: flex;
-      flex-grow: 1;
-
-      .input {
-        &.show-clear-button {
-          padding-right: 2rem;
-        }
-      }
-
-      .clear-button {
-        position: absolute;
-        inset: 0 0 0 auto;
-        font-size: @tiny-icon-size;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        padding: @light-spacing;
-        color: @text-muted;
-      }
-
-      &:not(:last-child) .input {
-        border-top-right-radius: 0;
-        border-bottom-right-radius: 0;
-        border-right: 0;
-      }
-    }
-
-    .button {
-      border-top-left-radius: 0;
-      border-bottom-left-radius: 0;
-    }
-  }
-
-  .char-limit {
-    float: right;
-    margin-right: @light-spacing;
-    font-size: @mini-text-size;
-    &:extend(.font-muted);
-    margin-left: auto;
-  }
-}
