--- components/interactables/InputGroup/types.d.ts
@@ -1,10 +1,0 @@
-export type InputType = 'text' | 'number' | 'password'
-export type InputColor =
-  | 'primary'
-  | 'link'
-  | 'info'
-  | 'success'
-  | 'warning'
-  | 'danger'
-  | null
-export type InputSize = 'sm' | 'md' | 'lg' | 'xl'
