+++ components/interactables/UserPicker/UserPicker.html
@@ -11,8 +11,7 @@
   </div>
   <InteractablesInput
     v-model="filter"
-    size="small"
-    type="dark"
+    size="xs"
     :placeholder="$t('servers.create.select_friends_placeholder')"
   />
   <div class="users" :style="{ height: height }">
