+++ components/views/chat/enhancers/Enhancers.html
@@ -47,7 +47,7 @@
               <InteractablesInput
                 v-model="search"
                 type="dark"
-                size="small"
+                size="xs"
                 :placeholder="$t('ui.search')"
               />
             </div>
