+++ components/views/chat/enhancers/glyphs/Glyphs.html
@@ -2,7 +2,7 @@
   <div class="search-glyph-box">
     <InteractablesInput
       :placeholder="$t('ui.search')"
-      size="small"
+      size="xs"
       v-model="searchText"
       input-kind="text"
       type="primary"
