+++ components/views/files/controls/Controls.html
@@ -6,9 +6,7 @@
       :tooltipPosition="'top'"
     >
       <InteractablesInput
-        size="small"
-        input-kind="text"
-        type="dark"
+        size="xs"
         :placeholder="$t('pages.files.browse.search')"
         class="search"
       />
@@ -32,7 +30,7 @@
     >
       <file-plus-icon size="1.3x" />
     </InteractablesButton>
-    <InteractablesInputGroup
+    <InteractablesInput
       v-model="text"
       class="name"
       size="xs"
@@ -40,7 +38,7 @@
       :placeholder="$t('pages.files.controls.name_folder')"
     >
       <folder-plus-icon size="1.3x" />
-    </InteractablesInputGroup>
+    </InteractablesInput>
     <input
       ref="upload"
       type="file"
