+++ components/views/marketplace/searchbox/Searchbox.html
@@ -1,9 +1,7 @@
 <div class="search-box">
   <InteractablesInput
     :placeholder="$t('ui.search')"
-    size="small"
+    size="xs"
     v-model="internalText"
-    input-kind="text"
-    type="primary"
   />
 </div>
