+++ components/views/navigation/server/sidebar/Sidebar.html
@@ -11,8 +11,7 @@
       <InteractablesInput
         :placeholder="`${$t('ui.search')}...`"
         size="small"
-        input-kind="text"
-        :delete-icon="true"
+        showClear
         type="dark"
         full-width
       />
