+++ components/views/navigation/sidebar/Sidebar.html
@@ -9,12 +9,9 @@
       >
         <InteractablesInput
           :placeholder="`${$t('ui.search')}...`"
-          size="small"
-          input-kind="text"
-          :delete-icon="true"
-          type="dark"
+          size="xs"
           disabled
-          full-width
+          showClear
         />
       </UiComingSoon>
       <InteractablesSidebarToggle />
