+++ components/views/servers/create/Create.html
@@ -39,8 +39,6 @@
         <InteractablesInput
           :placeholder="$t('servers.create.server_name_placeholder')"
           v-model="name"
-          input-kind="text"
-          type="primary"
         />
       </div>
