+++ components/views/settings/pages/privacy/Privacy.html
@@ -80,7 +80,7 @@
       <InteractablesInput
         v-model="ownInfo"
         type="dark"
-        size="small"
+        size="xs"
         :placeholder="$t('pages.privacy.ownInfo.placeholder')"
         class="ownInfo-input"
       />
