+++ components/views/wallet/balance/Balance.html
@@ -8,9 +8,9 @@
       <TypographyTitle :size="5" text="123,456.78" />
       <TypographyText> Last Updated: Today </TypographyText>
       <div class="inline">
-        <InteractablesInputGroup class="modal-input" size="xs">
+        <InteractablesInput class="modal-input" size="xs">
           <send-icon size="1x" />
-        </InteractablesInputGroup>
+        </InteractablesInput>
         <InteractablesSelect class="form-control" />
       </div>
       <div class="inline">
